backend/requirements.txt
fastapi
uvicorn
python-multipart

pdfplumber
python-docx
pytesseract
Pillow

scikit-learn
sentence-transformers
spacy

numpy
pandas
python-dateutil